Fix import functionality
